portfolioAPI/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
idna==2.7
pipenv==2018.11.26
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.4.0
